Conflicts:
	libs/scale/onboxtest/SysFsDiskOnBoxTest.cpp

+++ unittest/FStringUnitTest.cpp
@@ -266,7 +266,7 @@ TEST_F(FStringTest, ImplodeExplodeBinaryFStringTest)
 
     ASSERT_EQ(components.size(), componentsTest.size());
 
-    for (int i=0; i < components.size(); i++)
+    for (int i=0; i < (int)components.size(); i++)
     {
         ASSERT_STREQ( components[i], componentsTest[i] );
     }
